% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv.R
\name{BackFittingLocal}
\alias{BackFittingLocal}
\title{Estimación semiparamétrica mediante 'backfitting' local.}
\usage{
BackFittingLocal(
  frm.param,
  smooth.term = "s(x,bs=\\"cr\\",k=24)",
  cal.pts = NULL,
  datos,
  indice0,
  coords,
  var.fecha,
  var.loc = NULL,
  baseday = "2012-12-31",
  bw = 5000,
  bws = NULL,
  cores,
  tol = 0.001,
  plotind = FALSE
)
}
\arguments{
\item{frm.param}{Fórmula de la parte paramétrica (GWR) del modelo.}

\item{smooth.term}{Término no paramétrico del modelo.}

\item{cal.pts}{Puntos del espacio para los cuales se calcula el índice}

\item{datos}{Dataframe espacial conteniendo los datos.}

\item{indice0}{Valores iniciales del índice. Optativo: si está ausente, se calcula.}

\item{coords}{Coordenadas espaciales de las observaciones. Pueden ser geográficas (latitud y longitud) en cuyo caso se calcula distancia geodésica, o coordenadas de una proyección plana, en cuyo caso de emplea distancia euclídea ordinaria.}

\item{var.fecha}{Variable que recoge la fecha de cada observación. Debe formar parte de \code{datos.}}

\item{var.loc}{Variable que en el modelo base se emplea como proxy de la ubicación. Se elimina en el modelo que realiza GWR}

\item{baseday}{Día en que se fija la base 100 del índice.}

\item{bw}{"Bandwidth" o radio del área espacial que recibe ponderación sustancial en la estimación de parámetros de la GWR para cada punto de calibración.}

\item{bws}{"Bandwidth" o radio del área espacial que recibe ponderación sustancial en la estimación del spline}

\item{cores}{Número de cores que se desea emplear. Si no se indica nada, todos los de la máquina.}

\item{tol}{Máxima discrepancia relativa entre los valores de dos splines en iteraciones sucesivas para continuar la iteración.}

\item{plotind}{Variable lógica: TRUE si se desea la impresión del índice resultante.}
}
\value{
Una lista de índices de precios, con base en el día especificado, en formato de serie temporal \code{zoo;} uníndice para cada punto especificado en \code{cal.pts}.
}
\description{
Realiza la estimación de un modelo semi-paramétrico, cuya parte
paramétrica esta formada por una regresión espacialmente ponderada y
cuya parte no paramétrica está formada por un spline cúbico para cada
ubicación. Esto permite el ajuste de tendencias locales
}
\examples{
Ver viñeta.

}
