% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv.R
\name{cloromap}
\alias{cloromap}
\title{cloromap}
\usage{
cloromap(
  sp,
  var.sp,
  df,
  var.df,
  var,
  sub.var = NULL,
  sub.set = NULL,
  legend.title = NULL,
  logcolor = FALSE,
  graf.title = NULL,
  ...
)
}
\arguments{
\item{sp}{Spatial dataframe, in whose @data slot variable \code{var.sp} defines the aggregation zones.}

\item{var.sp}{Variable whose values specify the spatail aggregation.}

\item{df}{Ordinary dataframe with variable \code{var.df} matching \code{var.sp.}}

\item{var.df}{Variable matching \code{var.sp}; both need to have the same levels.}

\item{var}{Variable to code with color. Must be a column of \code{df.}}

\item{sub.var}{Variable in \code{sp}, optionally used to subset \code{sp}.}

\item{sub.set}{Vector of values of \code{sub.var} defining a subser of rows of \code{sp}.}

\item{legend.title}{Legend heading}

\item{logcolor}{FALSE if linear color scale, TRUE if logarithmic.}

\item{graf.title}{Heading of the graph.}

\item{...}{Optional additional parameters to pass to the underlying \code{ggplot} function.}
}
\value{
A map, of ggplot2 format, printed by default.
}
\description{
Creates cloropleth maps coding in color a variable.
}
\examples{

}
