% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv.R
\name{CompFechas}
\alias{CompFechas}
\title{CompFechas}
\usage{
CompFechas(indice, fechas = NULL)
}
\arguments{
\item{indice}{An index, such as returned by \code{ConsInd.}, of class \code{Indice} or \code{IndiceCB}}

\item{fechas}{Vector of dates whose extremes define the new date range. Can be omitted, in which case the extremes are taken from the index of \code{indice}}
}
\value{
A dily index over the period from \code{min(fechas)} to \code{max(fechas).}
}
\description{
Given an index, defined over a set of incomplete dates, interpolates a new index over the full set of dates.
}
\examples{

}
