% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv.R
\name{ConsInd}
\alias{ConsInd}
\title{ConsInd}
\usage{
ConsInd(
  modelo = NULL,
  conf = NULL,
  fechas = NULL,
  tit = tit,
  ylabel = "Índice",
  plot = FALSE,
  basedate = "2008-01-01",
  basevalue = 100
)
}
\arguments{
\item{modelo}{Modelo semiparamétrico estimado por \code{gam.}}

\item{conf}{Confianza del intervalo; si no se especifica, 95\%.}

\item{fechas}{Vector de fechas de las observaciones; habitualmente se toma de una columna de la dataframe empleada por \code{gams} para estimar \code{modelo}}

\item{tit}{Encabezamiento del gráfico producido con el índice.}

\item{ylabel}{Rótulo eje de ordenadas}

\item{plot}{Variable lógica: si es \code{FALSE} (default) no se genera el plot, en caso contrario sí.}

\item{base}{Fecha (día) tomada como base 100 del índice.}

\item{x.base}{Primera fecha en el plot; habitualmente en blanco y obtenida a partir de \code{fechas.}}
}
\value{
Un indice en formato \code{zoo}, que opcionalmente (plot=TRUE) es representado gráficamente.
}
\description{
Dado un modelo semi-paramétrico ya estimado, que incluye un término spline recogiendo la tendencia,
extrae dicho término, lo completa para los días en que no haya observación, y lo devuelve como una serie temporal, representándolo gráficamente
si se desea. Se supone que la variable respuesta el log(Precio) o log(Precio/m2); el índice devuelto lo es para la variable Precio o Precio/m2, respectivamente, y datos fechados diariamente.
}
\examples{

}
