% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_facets.R
\name{coord_facets}
\alias{coord_facets}
\title{Coord Facets}
\usage{
coord_facets(data, subradius, rotate = 0)
}
\arguments{
\item{data}{list generated by \code{\link{input_excel}}, or by hand,
containing formatted data.}

\item{subradius}{integer; radius of the circles representing facets (same
unit as center distances).}

\item{rotate}{integer between 0 and 2*pi; radian angle to rotate the chart
counter-clockwise by.}
}
\value{
List containing coordinates of chart objects.
}
\description{
Generates coordinates for a facet chart.
}
\details{
Use this function in conjunction with \code{\link{plot_facets}} for
  simple models.

  In a nested model, it is more efficient to use \code{\link{coord_nested}}
  instead. It automatically generates coordinates for all factors of the
  nested model.
}
\examples{
# creating a facet chart is a two step process, using this function and plot_facets:
coord <- coord_facets(DSSEI,subradius = .5)
DSSEI_facets <- plot_facets(coord, filename = "DSSEI_facets")
DSSEI_facets

}
\seealso{
\code{\link{plot_facets}}
}
