% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv_est.R
\name{cormat}
\alias{cormat}
\title{Cor(relation) Mat(rix)}
\usage{
cormat(fit)
}
\arguments{
\item{fit}{fitted lavaan model}
}
\value{
matrix; correlation matrix
}
\description{
Retrieve factor correlation matrix from lavaan model
}
