% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_subradius.R
\name{def_subradius}
\alias{def_subradius}
\title{Default Subradius}
\usage{
def_subradius(data)
}
\arguments{
\item{data}{list generated by \code{\link{input_excel}}, or by hand,
containing formatted data.}
}
\value{
integer, default subradius
}
\description{
Calculates a subradius for facet charts that fits the data.
}
