% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_chart.R
\name{facet_chart}
\alias{facet_chart}
\title{Facet Chart}
\usage{
facet_chart(data, subradius, rotate_radians = 0, rotate_degrees = 0,
  size = 1, file = "pdf", filename = NULL, filewidth = 10,
  fileheight = 10, colour = "black", fade = 90, font = "sans",
  tick = 0.1, cor_labels = TRUE, size_cor_labels = 1,
  size_test_label = 1, size_facet_labels = 1, width_axes = 1,
  width_circles = 1, width_tick = 1, size_tick_label = 1,
  size_center_dot = 1)
}
\arguments{
\item{data}{list generated by \code{\link{input_excel}}, or by hand,
containing formatted data.}

\item{subradius}{integer; (arbitrary) radius of the circles representing
facets (same unit as center distances), use to avoid circle overlap and
optimize appearance.}

\item{rotate_radians}{integer; radian angle to rotate the chart
counter-clockwise by; use fractions of pi (e.g. pi/2 = 90 degrees).}

\item{rotate_degrees}{integer; angle in degrees to rotate the chart
counter-clockwise by.}

\item{size}{integer; smartly scaled size of chart objects.}

\item{file}{character; file format to save. Supported formats are: "pdf"
(best quality with lowest file size),"png","jpeg". Suppress file output
using any other value; defaults to "pdf".}

\item{filename}{character string; name of the file the chart is written to;
filename extension added from file parameter like this: [filename].[file].}

\item{filewidth}{integer; width of the .pdf file; defaults to 10, dpi is
3000.}

\item{fileheight}{integer; height of the .pdf file; defaults to 10, dpi is
3000.}

\item{colour}{character; name of the accent colour.}

\item{fade}{integer; brightness of the gray tones between 0 (black) and 100
(white) in steps of 1; defaults to 90.}

\item{font}{character; text font; defaults to 'sans' (use extrafonts to
install additional fonts).}

\item{tick}{numeric; axis tick position; defaults to .1.}

\item{cor_labels}{logical; if \code{TRUE}, draws latent correlations between
facets as text.}

\item{size_cor_labels}{integer; font size of the latent correlations between
facets (relative to default).}

\item{size_test_label}{integer; font size of the test label (relative to
default).}

\item{size_facet_labels}{integer; font size of the facet labels (relative to
default).}

\item{width_axes}{integer; width of the radial axis lines indicating center
distances (relative to default).}

\item{width_circles}{integer; width of the circle outlines (relative to
default).}

\item{width_tick}{integer; width of the dotted axis tick line (relative to
default).}

\item{size_tick_label}{integer; font size of the axis tick label (relative to
default).}

\item{size_center_dot}{integer; size of the center dot marking the origin
(relative to default).}
}
\value{
Object of the class "ggplot" and optionally the same object saved as
  a file.
}
\description{
Creates a facet chart, showing the facets of a test.
}
\details{
This function creates the chart in one step, but does not return the
  coordinates from \code{\link{coord_facets}}. Alternatively use
  \code{\link{coord_facets}} and \code{\link{plot_facets}}. This allows you to
  make changes to or reuse the coordinates.
}
\examples{
# as simple as that
facet_chart(SMTQ, subradius = .6)

}
\seealso{
\code{\link{coord_facets}} \code{\link{plot_facets}}
}
