% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv_est.R
\name{floads}
\alias{floads}
\title{Load(ing)s}
\usage{
floads(fit, vars = NULL)
}
\arguments{
\item{fit}{fitted lavaan model}

\item{vars}{character; variables for which loadings should be extracted;
defaults to NULL, in which case all variables are considered}
}
\value{
numeric; vector of standardized factor loadings
}
\description{
Extract the standardized factor loadings from a fitted lavaan model.
}
