% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv.R
\name{Fusion}
\alias{Fusion}
\title{Fusion}
\usage{
Fusion(
  X,
  Y = NULL,
  varX,
  locX = 1,
  varY,
  locY = 2,
  canon,
  locCanon = NULL,
  transX = NULL,
  transY = NULL
)
}
\arguments{
\item{X}{First dataframe to merge}

\item{Y}{Second dataframe to merge}

\item{varX}{Variable acting as key in X}

\item{locX}{Column in \code{canon} against which to match values of column \code{varX} in \code{X}}

\item{varY}{Variable acting as key in Y}

\item{locY}{Column in \code{canon} against which to match values of column \code{varY} in \code{Y}}

\item{canon}{Correspondence between varX, varY and canonical value of the key (defaults to column 'Canon').}

\item{locCanon}{Column in \code{canon} holding the canonical names. Defaults to column named 'Canon' if one exists and the argument is not set on call.}

\item{transX}{Transformation to be applied to varX (if any)}

\item{transY}{Transformation to be applied to varY (if any)}
}
\value{
Data frame or merge of two data frames, with location names replaced by canonical names.
}
\description{
Merge two dataframes using  \code{varX} and \code{varY} as keys. Since they may not be exactly alike,
\code{canon} contains a dataframe which enables translation of either key
to a canonical form after (optionally) transliterating to a common character encoding.
}
\examples{
X     <- data.frame(a=c("Bilbao","San Sebastian","Vitoria","Teruel"),
                    b=c("A","B","B","D"), d=c(TRUE,TRUE,FALSE,TRUE))
X
Y     <- data.frame(e=c("Bilbao","Donostia","Vitoria/Gazteiz","Soria"),
                    g=c("F","G","H","J"), h=c(TRUE,FALSE,TRUE,TRUE))
Y
canon <- data.frame(X=c("Bilbao","San Sebastian","Vitoria"),
                    Y=c("Bilbao","Donostia","Vitoria/Gazteiz"),
                    Canon=c("Bilbao","Donostia","Vitoria"))
canon

Fusion(X, varX="a", locX=1, locCanon=3, canon=canon)

Fusion(X=Y, varX="e", locX=2, locCanon=3, canon=canon)

Fusion(X,Y,varX="a",varY="e",locX=1, locY=2, canon=canon)

canon <- data.frame(X=c("Bilbao","San Sebastian","Vitoria","Teruel","Soria"),
                    Y=c("Bilbao","Donostia","Vitoria/Gazteiz", "Teruel","Soria"),
                    Canon=c("Bilbao","Donostia","Vitoria","Teruel","Soria"))

Fusion(X, varX="a", locX=1, locCanon=3, canon=canon)

Fusion(X=Y, varX="e", locX=2, locCanon=3, canon=canon)

Fusion(X,Y,varX="a",varY="e",locX=1, locY=2, canon=canon)

canon <- data.frame(X=c("Bilbao","San Sebastian","Vitoria","Teruel","Soria"),
                    Y=c("Bilbao","Donostia","Vitoria/Gazteiz", "Teruel","Soria"),
                    Canon=c("CAPV","CAPV","CAPV","NoCAPV","NoCAPV"))

Fusion(X, varX="a", locX=1, locCanon=3, canon=canon)

Fusion(X=Y, varX="e", locX=2, locCanon=3, canon=canon)

# This hardly makes sense:

Fusion(X, Y, varX="a", varY="e", locX=1, locY=2, locCanon=3, canon=canon)

}
