% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv_est.R
\name{get_names}
\alias{get_names}
\title{Get names}
\usage{
get_names(dat)
}
\arguments{
\item{dat}{data frame; dataset}
}
\value{
data frame; names of tests, facets and items
}
\description{
Extract the names of tests, facets, and items from the variable names of a
dataset.
}
\details{
variable names in the data have to strictly match the following
  scheme: "test_facet_item" or "facet_item".
}
