% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv_est.R
\name{get_xarrows}
\alias{get_xarrows}
\title{Get Xarrows}
\usage{
get_xarrows(cors, design)
}
\arguments{
\item{cors}{list; list of latent correlation matrices of each model}

\item{design}{data frame; each facet (column "facet") is matched with its
superordinate test (column "test")}
}
\value{
data frame; data frame in the required format for the drawing of
  arrows in nested charts, including only those latent facet correlations,
  that exceed the correlation between the respective tests.
}
\description{
Creates a data frame for the drawing of arrows in nested charts, including
all correlations between facets that exceed the correlation of the respective
tests.
}
