% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HEXACO}
\alias{HEXACO}
\title{IPIP HEXACO Equivalent Scales}
\format{
An object of class \code{data.frame} with 20174 rows and 245 columns.
}
\usage{
HEXACO
}
\description{
Cleaned-up data from an ad-hoc online sample of n = 22786 participants on the
240 items of the IPIP HEXACO Equivalent Scales. Data were collected before
21st June 2014 within the openpsychometrics project
(\url{https://openpsychometrics.org/}). After listwise deletion of missing
values and including only those participants who did at least "agree" on the
items "I understand the instructions for this test." and "I have answered all
of these questions as accurately as possible.", data on n = 20174
participants remains.
}
\keyword{datasets}
