% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HEXACO_long}
\alias{HEXACO_long}
\title{IPIP HEXACO Equivalent Scales (long format)}
\format{
An object of class \code{data.frame} with 4887600 rows and 5 columns.
}
\usage{
HEXACO_long
}
\description{
See \code{\link{HEXACO}}
}
\keyword{datasets}
