% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv_est.R
\name{ind_lav}
\alias{ind_lav}
\title{ind lav}
\usage{
ind_lav(vars, indicators)
}
\arguments{
\item{vars}{character; variable names}

\item{indicators}{character; indicator names, may include unused indicators}
}
\value{
character; lavaan model syntax
}
\description{
create a lavaan model syntax based on a set of variable names and indicator
names that comprise these variable names
}
\details{
Indicator names have to include the variable names like this:
  "...variable_...". Variable names have to be unique and cannot be contained
  in one another like this: "variable_" and "ariable_"
}
