% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv.R
\name{IndZonas}
\alias{IndZonas}
\title{IndZonas}
\usage{
IndZonas(datos, zonas, frm = NULL, base = "2008-01-01", plot = FALSE)
}
\arguments{
\item{datos}{Dataframe de datos}

\item{zonas}{Nombre de una columna de \code{datos} que define las zonas para las que se desea calcular los índices.}

\item{frm}{Fórmula especificando el modelo base que quiere emplearse, en la forma esperada por \code{gam.} La fórmula es común a todos los índices computados por la función.}

\item{base}{Fecha base (índice=100). Es una fecha diaria en formato yyyy-mm-dd.}

\item{plot}{\code{FALSE} (default) si no se desean gráficos, \code{TRUE} en caso contrario.}
}
\value{
Una lista con los índices calculados, en el formato en que los devuelve \code{ConsInd}.
}
\description{
Estimación de índices globales por zonas; es unam mera función de utilidad que invoca reiteradamente  \code{ConsInd}.
}
\examples{
Ver viñeta

}
