% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_excel.R
\name{input_excel}
\alias{input_excel}
\title{Input Excel}
\usage{
input_excel(global = NULL, factors)
}
\arguments{
\item{global}{character; name of the excel file containing data on the overall general factor and the factors.}

\item{factors}{character; name(s) of the excel file(s) containing data on the factor(s) and its/their facets.}
}
\value{
list containing formatted data including center distances for \code{\link{model_items}}, \code{\link{model_facets}}, and \code{\link{model_nested}}.
}
\description{
Reads excel files containing factor loadings and latent correlations for IPV plots.
}
\details{
Note that the excel files need a very specific structure. Use the example files below as templates.

The \code{global} argument defaults to NULL.
This allows to only use the \code{factors} argument, resulting in a simple model with one factor.
}
\examples{
# read data for a single factor by ignoring the "global" parameter of input_excel
single_file <- system.file("extdata", "DSSEI.xlsx", package = "IPV", mustWork = TRUE)
x <- input_excel(factors = single_file)

# read data for a nested factor model
# note that excel files need to be split as in the example to contain only
# one general factor model and one correlated factor model for the same set of items
global <- system.file("extdata", "IPV_global.xlsx", package = "IPV", mustWork = TRUE)
factors <- c(system.file("extdata", "IPV_DSSEI.xlsx", package = "IPV", mustWork = TRUE),
             system.file("extdata", "IPV_SMTQ.xlsx", package = "IPV", mustWork = TRUE),
             system.file("extdata", "IPV_RSES.xlsx", package = "IPV", mustWork = TRUE))
x <- input_excel(global = global,factors = factors)

}
