% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_factor_excel.R
\name{input_factor_excel}
\alias{input_factor_excel}
\title{Input Factor Excel}
\usage{
input_factor_excel(file)
}
\arguments{
\item{file}{character; filename of the excel file}
}
\value{
list containing formatted data including center distances for use with \code{\link{model_items}}, \code{\link{model_facets}}
}
\description{
This function reads factor loadings and latent correlations for factor and its facets from a single excel file.
}
\details{
Do not use this function, always use \code{\link{input_excel}} instead.
}
\seealso{
\code{\link{input_excel}}
}
