% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_manual_process.R
\name{input_manual_process_factor}
\alias{input_manual_process_factor}
\title{Input Manual Process Factor}
\usage{
input_manual_process_factor(data)
}
\arguments{
\item{data}{list generated by \code{\link{input_manual_simple}} with complete
data.}
}
\value{
List containing formatted data including center distances for a
  single factor.
}
\description{
Helper function of \code{\link{input_manual_process}}.
}
