% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPV.R
\docType{package}
\name{IPV}
\alias{IPV}
\alias{IPV-package}
\title{IPV: A package to create Item Pool Visualizations}
\description{
The IPV package provides three categories of important functions:
input, model, and plot.
}
\section{Input Functions}{


The input functions are used to format your raw data, so the model
functions can use them.
}

\section{Model Functions}{


The model functions are used to calculate specific coordinates the plot
functions draw from.
}

\section{Plot Functions}{


The plot functions are used to create and customize the actual plots.
}

