% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv_expand.R
\name{ipv_expand}
\alias{ipv_expand}
\title{IPV expand}
\usage{
ipv_expand(est, est_raw)
}
\arguments{
\item{est}{list; estimates including center distances}

\item{est_raw}{list; raw estimates}
}
\value{
object of class "IPV"
}
\description{
Helper function to expand estimates of simple IPV to full sized object of
class "IPV"
}
