% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv_est.R
\name{ipv_long_to_wide}
\alias{ipv_long_to_wide}
\title{IPV long to wide}
\usage{
ipv_long_to_wide(x, id = "id", value.var = "value")
}
\arguments{
\item{x}{data frame; raw data in long format}

\item{id}{character; name of case identifying variable}

\item{value.var}{character; name of variable that contains measurement values}
}
\value{

}
\description{
Helper function to convert long format data into appropriate wide format for
ipv_est
}
