% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_chart.R
\name{item_chart}
\alias{item_chart}
\title{Item Chart}
\usage{
item_chart(data, filename = "myipv.pdf", size = 1, font = "sans",
  rotate_radians = 0, rotate_degrees = 0, filewidth = 12,
  fileheight = 10, dpi = 500, color = "black", color2 = "black",
  fade = 0, dodge = 1, rotate_title_radians = 0,
  rotate_title_degrees = 0, width_items = 1, length_items = 1,
  length_ratio_items = 1.5, size_tick_label = 1, size_title = 1,
  size_facet_labels = 1, width_axes = 1, size_arrow_heads = 1,
  width_grid = 1)
}
\arguments{
\item{data}{list generated by \code{\link{input_excel}}, or by hand,
containing formatted data.}

\item{filename}{character; name of the file to save. Supported formats are:
"pdf" (best quality with lowest file size), "png", "jpeg". Use "none" to
suppress file output; defaults to "myipv.pdf".}

\item{size}{integer; size of all chart objects except item bars.}

\item{font}{character; text font; defaults to 'sans' (use extrafonts to
install additional fonts).}

\item{rotate_radians}{integer; radian angle to rotate the chart
counter-clockwise by; use fractions of pi (e.g. pi/2 = 90 degrees).}

\item{rotate_degrees}{integer; angle in degrees to rotate the chart
counter-clockwise by.}

\item{filewidth}{integer; width of the graphic in inches; defaults to 12.}

\item{fileheight}{integer; height of the graphic in inches; defaults to 10.}

\item{dpi}{integer; resolution in dots per inch; defaults to 500.}

\item{color}{character; name of the first accent color; defaults to black.}

\item{color2}{character; name of the second accent color; defaults to black.}

\item{fade}{integer; brightness of grid lines, axes, and center dot; values
can range from -4 (almost white) to 2.6 (almost black) in steps of .2;
defaults to 0.}

\item{dodge}{integer; relative amount of horizontal outward dodge of
facet labels.}

\item{rotate_title_radians}{integer; radian angle to rotate the global label
counter-clockwise by; use fractions of pi (e.g. pi/2 = 90 degrees).}

\item{rotate_title_degrees}{integer; angle in degrees to rotate the global
label counter-clockwise by.}

\item{width_items}{integer; width of the item bars (relative to default).}

\item{length_items}{integer; length of the item bars (relative to default).}

\item{length_ratio_items}{integer; relative length of every other item bar;
defaults to 1.5.}

\item{size_tick_label}{integer; font size of the axis tick label (relative to
default).}

\item{size_title}{integer; font size of the test label (relative to default).}

\item{size_facet_labels}{integer; font size of the axis labels (relative to
default).}

\item{width_axes}{integer; width of the radial axis lines indicating center
distances (relative to default).}

\item{size_arrow_heads}{integer; length of the arrow heads at the end of the
axes (relative to default).}

\item{width_grid}{integer; width of the dotted grid lines (relative to
default).}
}
\value{
Object of the class "ggplot" and optionally the same object saved as
  a file.
}
\description{
Creates an item chart, showing the items of a test arranged by facets.
}
\details{
When changing the size of objects, consider the \code{size}
  parameter first and make specific adjustments with the other \code{size_}
  and \code{width_} parameters after.

  To better display overlapping item values, change the width of the item
  bars, or set the accent colors to different values, or change the ratio of
  item lengths.
}
\examples{
# as simple as that
item_chart(SMTQ)

}
\seealso{
\code{\link{facet_chart}} \code{\link{nested_chart}}
}
