% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv.R
\name{mggplot}
\alias{mggplot}
\title{mggplot}
\usage{
mggplot(
  x,
  columnas = names(x),
  tipo = c("panel", "multiple"),
  leyenda = c("bottom", "right")
)
}
\arguments{
\item{x}{Lista de índices tal como los proporciona \code{ConsInd.} Puede ser una dataframe  (caso especial de lista).}

\item{columnas}{Nombres de los índices a representar; por defecto, los nombres de la lista \code{x} (o columnas de la dataframe \code{x}.)}

\item{tipo}{Tipo de plot que se desea: "panel" si se desea cada serie en un panel propio, o "multiple", si se desean todas las series en un único panel.}

\item{leyenda}{Lugar donde situar la leyenda; por defecto, "bottom", pero puede especificarfse "right".}
}
\value{
Llamada por su efecto secundario, consistente en generar los gráficos.
}
\description{
Representar múltiples índices (típicamente devueltos por la función \code{IndZonas}) en sendos paneles o superpuestos en un único panel.
}
\examples{

}
