% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_facets.R
\name{model_facets}
\alias{model_facets}
\title{Model Facets}
\usage{
model_facets(data, subradius, rotate = 0)
}
\arguments{
\item{data}{list generated by \code{\link{input_excel}}, or by hand, containing formatted data.}

\item{subradius}{integer; radius of the circles representing facets (same unit as center distances).}

\item{rotate}{integer between 0 and 2*pi; radian angle to rotate the plot counter-clockwise by.}
}
\value{
List containing coordinates of plot objects.
}
\description{
Generates coordinates for a facet plot.
}
\details{
Use this function in conjunction with \code{\link{plot_facets}} for simple models.

In a nested model, it is more efficient to use \code{\link{model_nested}} instead.
It automatically generates plot coordinates for all factors of the nested model.
}
\examples{
# creating plots is a two step process, using model_ and plot_ functions:
coord <- model_facets(DSSEI,subradius = .5)
DSSEI_item_plot <- plot_facets(coord,filename = "DSSEI_facets")

}
\seealso{
\code{\link{plot_facets}}
}
