% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_items.R
\name{model_items}
\alias{model_items}
\title{Model Items}
\usage{
model_items(data, rotate = 0)
}
\arguments{
\item{data}{list generated by \code{\link{input_excel}}, or by hand, containing formatted data.}

\item{rotate}{integer between 0 and 2*pi; radian angle to rotate the plot counter-clockwise by.}
}
\value{
List containing coordinates of plot objects.
}
\description{
Generates coordinates for an item plot.
}
\details{
Use this function in conjunction with \code{\link{plot_items}} for simple models.

In a nested model, it is more efficient to use \code{\link{model_nested}} instead.
Set \code{items = TRUE} to generate the item plot coordinates for all factors of the nested model.
}
\examples{
# creating plots is a two step process, using model_ and plot_ functions:
coord <- model_items(DSSEI)
DSSEI_item_plot <- plot_items(coord,filename = "DSSEI_items")

}
\seealso{
\code{\link{plot_items}} \code{\link{model_nested}}
}
