% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_nested.R
\name{model_nested}
\alias{model_nested}
\title{Model Nested}
\usage{
model_nested(data, subradius, rotate = 0, subrotate = 0, rotate_title = 0,
  items = FALSE, cors = TRUE, cor_spacing = 0.4, relative_scaling = 3,
  extra_arrows = NULL)
}
\arguments{
\item{data}{list generated by \code{\link{input_excel}}, or by hand, containing formatted data.}

\item{subradius}{integer; radius of the circles representing facets (same unit as center distances).}

\item{rotate}{integer between 0 and 2*pi; radian angle to rotate the plot counter-clockwise by.}

\item{subrotate}{integer between 0 and 2*pi; radian angle or vector of radian angles to rotate the nested plots counter-clockwise by.}

\item{rotate_title}{integer between 0 and 2*pi; radian angle to rotate the overall general factor label counter-clockwise by.}

\item{items}{logical; if \code{TRUE}, generates the coordinates for item plots for all factors by calling \code{\link{model_items}}.}

\item{cors}{logical; if \code{TRUE}, generates coordinates for latent correlations between factors and sets up a ring to plot them in. If \code{FALSE}, the ring and the correlations are omitted, simplifying the plot significantly.}

\item{cor_spacing}{integer; if \code{cors = TRUE}: width of the ring the latent correlations between factors are plotted in; defaults to 0.4.}

\item{relative_scaling}{integer; relative size of the global to the nested plots; defaults to 3.}

\item{extra_arrows}{data frame containing information about additional correlation arrows between facets of different factors; see examples.}
}
\value{
List containing coordinates of plot objects.
}
\description{
Generates coordinates for a nested plot and all other plots.
}
\details{
Use this function in conjunction with \code{\link{plot_nested}} for nested models.

This function also provides data for facet plots and item plots (if \code{items = TRUE})
In this case, use it in conjunction with \code{\link{plot_facets}} or \code{\link{plot_items}}.

If you set \code{subrotate} to a single value, all nested plots will be rotated by the same amount.
If you set it to a vector of values, the nested plots will be rotated by their specific values given in that vector.

Increase \code{relative_scaling} to avoid circle overlap. Decrease it to make small plot objects more visible.

\code{cors} and \code{cor_spacing} add larger circles around the nested facet plots, but do not change them.
}
\examples{
# creating plots is a two step process, using model_ and plot_ functions:
coord <- model_nested(self_confidence,subradius = .6)
sc_plot <- plot_nested(coord,filename = "self_confidence_nested")

# adding extra arrows
sc_arrows <- data.frame(V1_factor=rep(NA,3),
                        V1_subfactor=rep(NA,3),
                        V2_factor=rep(NA,3),
                        V2_subfactor=rep(NA,3),
                        value=rep(NA,3))
sc_arrows[1,] <- c("DSSEI","Ab","RSES","Ps",".67")
sc_arrows[2,] <- c("DSSEI","Ab","SMTQ","Cs",".81")
sc_arrows[3,] <- c("SMTQ","Ct","RSES","Ns",".76")
coord <- model_nested(self_confidence,subradius = .6,extra_arrows = sc_arrows)
sc_plot <- plot_nested(coord,filename = "self_confidence_nested",extra_arrows = TRUE)

# rotating the nested facet plots one by one (compare to subrotate = 0)
coord <- model_nested(self_confidence,subradius = .6,subrotate = c(0,pi/2,0))
sc_plot <- plot_nested(coord,filename = "self_confidence_nested")

}
\seealso{
\code{\link{plot_nested}}
}
