% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_IPV.R
\name{new_IPV}
\alias{new_IPV}
\title{New IPV}
\usage{
new_IPV(xarrow = TRUE)
}
\arguments{
\item{xarrow}{logical; should the element xarrow be included?; defaults to TRUE}
}
\value{
IPV; empty IPV structure
}
\description{
Create empty structure of class IPV
}
