% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_facets.R
\name{plot_facets}
\alias{plot_facets}
\title{Plot Facets}
\usage{
plot_facets(coord, size = 1, filename = "myipv.pdf", filewidth = 10,
  fileheight = 10, dpi = 500, color = "black", fade = 85,
  font = "sans", cor_labels = TRUE, size_cor_labels = 1,
  size_test_label = 1, size_facet_labels = 1, width_axes = 1,
  width_circles = 1, width_tick = 1, size_tick_label = 1)
}
\arguments{
\item{coord}{list generated by \code{\link{coord_facets}} or
\code{\link{coord_nested}}.}

\item{size}{integer; simultaneously scales the size of most chart objects.}

\item{filename}{character; name of the file to save. Supported formats are:
"pdf" (highest quality and smallest file size), "png", "jpeg". Use "none"
to suppress file output; defaults to "myipv.pdf".}

\item{filewidth}{integer; width of the .pdf file in inches; defaults to 10.}

\item{fileheight}{integer; height of the .pdf file in inches; defaults to 10.}

\item{dpi}{integer; resolution in dots per inch for "png" and "jpeg" files;
defaults to 500.}

\item{color}{character; name of the accent color; defaults to "black".}

\item{fade}{integer; brightness of the gray tones between 0 = "black" and 100
= "white" in steps of 1; defaults to 85.}

\item{font}{character; text font, use extrafonts to access additional fonts;
defaults to "sans", which is "Helvetica".}

\item{cor_labels}{logical; if \code{TRUE}, shows latent correlations between
facets; defaults to \code{TRUE}.}

\item{size_cor_labels}{integer; correlation font size relative to default.}

\item{size_test_label}{integer; test font size relative to default.}

\item{size_facet_labels}{integer; facet font size relative to default.}

\item{width_axes}{integer; radial axis width relative to default.}

\item{width_circles}{integer; facet circle outline width relative to default.}

\item{width_tick}{integer; axis tick line width relative to default.}

\item{size_tick_label}{integer; axis tick font size relative to default.}
}
\value{
Object of the class "ggplot" and, by default, the same object saved
  as a file.
}
\description{
Generates a facet chart from coordinates.
}
\details{
Use \code{\link{facet_chart}} to create facet charts.
}
\seealso{
\code{\link{coord_facets}} \code{\link{facet_chart}}
}
