% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_items.R
\name{plot_items}
\alias{plot_items}
\title{Plot Items}
\usage{
plot_items(coor, size = 1, filename, width = 10, height = 10,
  font = "sans", colour = "black", colour2 = "black", tick_label = TRUE,
  size_title = 1, size_axis_labels = 1, width_axes = 1,
  size_arrow_heads = 1, width_items = 1, size_grid = 1,
  size_tick_label = 1, size_center_dot = 1)
}
\arguments{
\item{coor}{list generated by \code{\link{coord_items}} or
\code{\link{coord_nested}}.}

\item{size}{integer; smartly scaled size of chart objects.}

\item{filename}{character string; name of the .pdf file the chart is written
to; filename extension added automatically.}

\item{width}{integer; width of the .pdf document; defaults to 10; dpi is
3000.}

\item{height}{integer; height of the .pdf document; defaults to 10, dpi is
3000.}

\item{font}{character; text font; defaults to 'sans' (use extrafonts to
install additional fonts).}

\item{colour}{character; name of the first accent colour.}

\item{colour2}{character; name of the second accent colour.}

\item{tick_label}{logical; if \code{TRUE}, draws a text label for the axis
tick}

\item{size_title}{integer; font size of the test label (relative to
default).}

\item{size_axis_labels}{integer; font size of the axis labels (relative to
default).}

\item{width_axes}{integer; width of the radial axis lines (relative to
default).}

\item{size_arrow_heads}{integer; length of the arrow heads (relative to
default).}

\item{width_items}{integer; width of the item bars (relative to default).
Currently not supported!}

\item{size_grid}{integer; width of the dotted grid lines (relative to
default).}

\item{size_tick_label}{integer; font size of the axis tick label (relative to
default).}

\item{size_center_dot}{integer; size of the center dot marking the origin
(relative to default).}
}
\value{
Object of the class "ggplot" and the same object as a .pdf file.
}
\description{
Plots an item chart, showing the items of a test arranged by facets.
}
\details{
Use this function in conjunction with \code{\link{coord_items}} or
  \code{\link{coord_nested}} for simple models.

  To better display overlapping item values, set the accent colours to
  different values.
}
\examples{
# creating item charts is a two step process, using coord_items and this function:
coord <- coord_items(DSSEI)
DSSEI_items <- plot_items(coord,filename = "DSSEI_items")
DSSEI_items

}
\seealso{
\code{\link{coord_items}}
}
