% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv.R
\name{pols}
\alias{pols}
\title{pols}
\usage{
pols(base, sub.var = NULL, sub.set = NULL, var.agreg = NULL)
}
\arguments{
\item{base}{Dataframe espacial}

\item{sub.var}{Variable a partir de la cual seleccionaremos un subconjunto, si procede.}

\item{sub.set}{Niveles de \code{sub.var} definiendo el subconjunto que queremos.}

\item{var.agreg}{Variable de agregación: se agruparán todos los polígonos que tengan mismos nivel de esta variable.}
}
\value{
Una dataframe espacial de polígonos para los ámbitos agregados.
}
\description{
Generación de poligonos por agrupación de otros. Por ejemplo, podemos tener una
dataframe espacial (\code{base}) con polígonos de municipios y querer otra con polígonos de
comarcas o áreas funcionales. Entonces, en \code{var.agreg} daríamos el nombre de la
columna que nombra las comarcas o áreas funcionales
y obtendríamos una dataframe espacial con los contornos de las mismas.
}
\details{
Podemos especificar que queremos solo un fragmento indicado una variable (\code{sub.var})
y los niveles de la misma a que deseamos limitarnos (\code{sub.set}). Por ejemplo,
\code{sub.var} podría ser \code{PROVINCIA} y \code{sub.set} \code{VIZCAYA}. Aunque la dataframe
original contuviera los tres territorios de la CAPV, el resltado contendría la agregación
por comarcas o áreas funcionales de sólo Bizkaia.
}
\examples{
Ver viñeta

}
