% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{relabel_est}
\alias{relabel_est}
\title{Relabel estimates}
\usage{
relabel_est(data, before, after)
}
\arguments{
\item{data}{IPV chart creation data (nested or simple)}

\item{before}{character; a vector of names to replace}

\item{after}{character; a vector of replacement names}
}
\value{
the same data with renamed values / variables
}
\description{
Relabel tests, facets, or items in IPV data
}
