% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{relabel_xarrow}
\alias{relabel_xarrow}
\title{Relabel Xarrow}
\usage{
relabel_xarrow(data, before, after)
}
\arguments{
\item{data}{IPV estimates for extra arrows (as provided e.g. by
\code{ipv_est})}

\item{before}{character; a vector of names to replace}

\item{after}{character; a vector of replacement names}
}
\value{
the same data with renamed values / variables
}
\description{
Relabel Xarrow
}
