% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\title{Rename}
\usage{
rename(data, before, after)
}
\arguments{
\item{data}{IPV chart creation data (nested or simple)}

\item{before}{character; a vector of names to replace}

\item{after}{character; a vector of replacement names}
}
\value{
the same data with renamed values / variables
}
\description{
Renames tests, facets, or items in IPV data
}
\examples{
rename(self_confidence,
c("RSES", "Ct", "7"),
c("Rosenberg", "control", "item_7"))

}
