% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_est}
\alias{rename_est}
\title{Rename estimates}
\usage{
rename_est(data, before, after)
}
\arguments{
\item{data}{IPV chart creation data (nested or simple)}

\item{before}{character; a vector of names to replace}

\item{after}{character; a vector of replacement names}
}
\value{
the same data with renamed values / variables
}
\description{
Renames tests, facets, or items in IPV data
}
