% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_simple}
\alias{rename_simple}
\title{Rename Simple}
\usage{
rename_simple(data, before, after, regex = FALSE)
}
\arguments{
\item{data}{IPV chart creation data (nested or simple)}

\item{before}{character; a vector of names to replace}

\item{after}{character; a vector of replacement names}

\item{regex}{logical; should items be renamed based on regular expressions
for treatment of global section in nested data?; defaults to FALSE}
}
\value{
the same data with renamed values / variables
}
\description{
Rename Simple
}
\details{
This function does not support regular expressions. Provide full
  names only.
}
