% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_xarrow}
\alias{rename_xarrow}
\title{Rename Xarrow}
\usage{
rename_xarrow(data, before, after)
}
\arguments{
\item{data}{IPV estimates for extra arrows (as provided e.g. by \code{ipv_est})}

\item{before}{character; a vector of names to replace}

\item{after}{character; a vector of replacement names}
}
\value{
the same data with renamed values / variables
}
\description{
Rename Xarrow
}
