% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{self_confidence}
\alias{self_confidence}
\title{Self-Confidence}
\format{An object of class \code{list} of length 2.}
\source{
Dantlgraber, M., Stieger, S., & Reips, U.-D. (in revision).
  Introducing Item Pool Visualization (IPV)
}
\usage{
self_confidence
}
\description{
A set of center distances and latent correlations for items and facets of the
Domain Specific Self-Esteem Inventory (DSSEI), Sports Mental Toughness
Questionnaire (SMTQ), and Rosenberg Self-Esteem Scale (RSES) completed by
2272 german speaking participants using the german versions of the
questionnaires.
}
\keyword{datasets}
