% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_factor.R
\name{shift_factor}
\alias{shift_factor}
\title{Shift Factor}
\usage{
shift_factor(data, x, y)
}
\arguments{
\item{data}{list generated by \code{\link{model_facets}} containing the coordinates of the parts of the plot.}

\item{x}{integer; horizontal shift.}

\item{y}{integer; vertical shift.}
}
\value{
A list of dataframes containing a limited selection of coordinates of plot objects.
}
\description{
This function moves the coordinates given by \code{\link{model_facets}} to nest the facet plots in \code{\link{model_nested}}.
}
