% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SMTQ}
\alias{SMTQ}
\title{Sports Mental Toughness Questionnaire}
\format{An object of class \code{list} of length 3.}
\source{
Dantlgraber, M., Stieger, S., & Reips, U.-D. (in revision).
  Introducing Item Pool Visualization (IPV)
}
\usage{
SMTQ
}
\description{
A set of center distances and latent correlations for items and facets of the
SMTQ, completed by 2272 german speaking participants using the german
version of the questionnaire.
}
\keyword{datasets}
