% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv_est.R
\name{write_IPV_syntax}
\alias{write_IPV_syntax}
\title{Write IPV syntax}
\usage{
write_IPV_syntax(dat, name)
}
\arguments{
\item{dat}{data frame; correctly formatted raw data}

\item{name}{character; name of the overall construct or test}
}
\value{
list of character; lavaan model syntax
}
\description{
Write lavaan model syntax of IPV models on the given dataset
}
\details{
Variable names in dat have to conform to the pattern
  "test_facet_item". If there is only one test in the data, the pattern is
  "facet_item". For tests without facets in a larger dataset also comprising
  tests with items, the pattern is "test_item". Variable names have to be
  unique at the level of the test AND the facet. Item names have to be unique
  at the level of the test (not only at the level of the facet).
}
