% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv.R
\name{BackFitting}
\alias{BackFitting}
\title{Semi-parametric estimation usign backfiffing.}
\usage{
BackFitting(
  frm.param,
  smooth.term = "s(x,bs=\\"cr\\",k=24)",
  var.loc = NULL,
  datos,
  indice0,
  coords,
  var.fecha,
  baseday = "2012-12-31",
  bw = 5000,
  cores,
  tol = 0.001,
  plotind = FALSE
)
}
\arguments{
\item{frm.param}{Formula for the parametric part  (GWR) of the model}

\item{smooth.term}{Non-parametric part of the model}

\item{var.loc}{Variable que en el modelo base se emplea como proxy de la ubicación. Se elimina en el modelo que realiza GWR}

\item{datos}{Dataframe espacial conteniendo los datos.}

\item{indice0}{Valores iniciales del índice. Optativo: si está ausente, se calcula.}

\item{coords}{Coordenadas espaciales de las observaciones. Pueden ser geográficas (latitud y longitud) en cuyo caso se calcula distancia geodésica, o coordenadas de una proyección plana, en cuyo caso de emplea distancia euclídea ordinaria.}

\item{var.fecha}{Variable que recoge la fecha de cada observación. Debe formar parte de \code{datos.}}

\item{baseday}{Día en que se fija la base 100 del índice.}

\item{bw}{"Bandwidth" o radio del área espacial que recibe ponderación sustancial en la estimación de parámetros para cada punto de calibración.}

\item{cores}{Número de cores que se desea emplear. Si no se indica nada, todos los de la máquina.}

\item{tol}{Máxima discrepancia relativa entre los valores de dos splines en iteraciones sucesivas para continuar la iteración.}

\item{plotind}{Variable lógica: TRUE si se desea la impresión del índice resultante.}
}
\value{
Un índice de precios, con base en el día especificado, en formato de serie temporal \code{zoo.}
}
\description{
This function fits a parametric model whose parametric part is
a geographically weigthed regression (GWR) while the non-parametric
is a cubic spline global, to all observations
}
\examples{
Ver viñeta.

}
