% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv.R
\name{cloromap}
\alias{cloromap}
\title{cloromap}
\usage{
cloromap(
  sp,
  var.sp,
  df,
  var.df,
  var,
  sub.var = NULL,
  sub.set = NULL,
  legend.title = NULL,
  logcolor = FALSE,
  graf.title = NULL,
  ...
)
}
\arguments{
\item{sp}{Dataframe espacial, en cuyo slot @data está la variable \code{var.sp} definiendo los ámbitos sobre los que se agrega.}

\item{var.sp}{Variable proporcionando la desagregación espacial.}

\item{df}{Dataframe ordinaria, conteniendo una variable \code{var.df} emparejable con \code{var.sp.}}

\item{var.df}{Variable emparejable con \code{var.sp}; ambas han de tener el mismo número de niveles e idénticamente nombrados.}

\item{var}{Variable a representar. Debe ser una columna de \code{df.}}

\item{sub.var}{Variable de \code{sp} utilizada para definir un subconjunto de filas de la misma.}

\item{sub.set}{Vector de valores de \code{sub.var} que definen un subconjunto de filas de \code{sp}.}

\item{legend.title}{Encabezamiento de la leyenda.}

\item{logcolor}{FALSE si la escala de color es lineal, TRUE si logarítmica}

\item{graf.title}{Encabedzamiento del gráfico.}

\item{...}{Parámetros adicionales a pasar a la función \code{ggplot} subyacente.}
}
\value{
Un mapa, en formato ggplot2, que por defecto se imprime.
}
\description{
Crea mapas codificando una variable de interés por zona con diferentes colores
}
\examples{
Ver viñeta

}
