% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv.R
\name{CompFechas}
\alias{CompFechas}
\title{CompFechas}
\usage{
CompFechas(indice, fechas)
}
\arguments{
\item{indice}{Un índice tal como lo devuelve \code{ConsInd.}}

\item{fechas}{Vector de fechas cuyos extremos definirán el nuevo fechado del índice}
}
\value{
Un índice diario, sobre todos los días que van de \code{min(fechas)} a \code{max(fechas).}
}
\description{
Dado un índice, definido sobre un conjunto de abscisas temporales con "huecos", define un nuevo índice interpolando los periodos sin observaciones.
}
\examples{

}
