% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_items.R
\name{coord_items}
\alias{coord_items}
\title{Coord Items}
\usage{
coord_items(data, rotate = 0)
}
\arguments{
\item{data}{list generated by \code{\link{input_excel}}, or by hand,
containing formatted data.}

\item{rotate}{integer between 0 and 2*pi; radian angle to rotate the chart
counter-clockwise by.}
}
\value{
List containing coordinates of chart objects.
}
\description{
Generates coordinates for an item chart.
}
\details{
Use this function in conjunction with \code{\link{plot_items}} for
  simple models.

  In a nested model, it is more efficient to use \code{\link{coord_nested}}
  instead. Set \code{items = TRUE} to generate item chart coordinates for
  all factors of the nested model.
}
\examples{
# creating an item chart is a two step process, using this function and plot_items:
coord <- coord_items(DSSEI)
DSSEI_items <- plot_items(coord, filename = "DSSEI_items")
DSSEI_items

}
\seealso{
\code{\link{plot_items}} \code{\link{coord_nested}}
}
