% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_items.R
\name{coord_items}
\alias{coord_items}
\title{Coord Items}
\usage{
coord_items(data, rotate_radians = 0, rotate_degrees = 0,
  rotate_title_radians = 0, rotate_title_degrees = 0, dodge_axes = 1)
}
\arguments{
\item{data}{list generated by \code{\link{input_excel}}, or by hand,
containing formatted data.}

\item{rotate_radians}{integer; radian angle to rotate the chart
counter-clockwise by; use fractions of pi (e.g. pi/2 = 90 degrees).}

\item{rotate_degrees}{integer; angle in degrees to rotate the chart
counter-clockwise by.}

\item{rotate_title_radians}{integer; radian angle to rotate the global label
counter-clockwise by; use fractions of pi (e.g. pi/2 = 90 degrees).}

\item{rotate_title_degrees}{integer; angle in degrees to rotate the global
label counter-clockwise by.}

\item{dodge_axes}{integer; relative amount of horizontal outward dodge of
axis labels.}
}
\value{
List containing coordinates of chart objects.
}
\description{
Generates coordinates for an item chart.
}
\details{
Use this function in conjunction with \code{\link{plot_items}} for
  simple models.

  Use \code{\link{item_chart}} to create the chart in a single step.

  Rotate using this function, not \code{\link{plot_items}}.

  In a nested model, it is more efficient to use \code{\link{coord_nested}}
  instead. Set \code{prepare_item_charts = TRUE} to generate item chart
  coordinates for all factors of the nested model.
}
\examples{
# # creating an item chart is a two step process, using this function and
# # plot_items:
# coord <- coord_items(DSSEI)
# DSSEI_items <- plot_items(coord, filename = "DSSEI_items")
# DSSEI_items
}
\seealso{
\code{\link{plot_items}} \code{\link{coord_nested}}
  \code{\link{item_chart}}
}
