% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_nested.R
\name{coord_nested}
\alias{coord_nested}
\title{Coord Nested}
\usage{
coord_nested(data, subradius, rotate_radians = 0, rotate_degrees = 0,
  subrotate_radians = 0, subrotate_degrees = 0, rotate_title_radians = 0,
  rotate_title_degrees = 0, prepare_item_charts = FALSE,
  correlations = TRUE, cor_spacing = 0.4, relative_scaling = 3,
  xarrows = NULL)
}
\arguments{
\item{data}{list generated by \code{\link{input_excel}}, or by hand,
containing formatted data.}

\item{subradius}{integer; (arbitrary) radius of the circles representing
facets (same unit as center distances), use to avoid circle overlap and
optimize appearance.}

\item{rotate_radians}{integer; radian angle to rotate the chart
counter-clockwise by; use fractions of pi (e.g. pi/2 = 90 degrees).}

\item{rotate_degrees}{integer; angle in degrees to rotate the chart
counter-clockwise by.}

\item{subrotate_radians}{integer; radian angle or vector of radian angles to
rotate the nested facet charts counter-clockwise by; use fractions of pi
(e.g. pi/2 = 90 degrees).}

\item{subrotate_degrees}{integer; angle in degrees or vector of angles in
degrees to rotate the nested facet charts counter-clockwise by.}

\item{rotate_title_radians}{integer; radian angle to rotate the global label
counter-clockwise by; use fractions of pi (e.g. pi/2 = 90 degrees).}

\item{rotate_title_degrees}{integer; angle in degrees to rotate the global
label counter-clockwise by.}

\item{prepare_item_charts}{logical; if \code{TRUE}, generates coordinates for
item charts for all factors by calling \code{\link{coord_items}}.}

\item{correlations}{logical; if \code{TRUE}, generates coordinates for latent
correlations between tests and sets up a ring to draw them in. If
\code{FALSE}, the ring and the correlations are omitted, simplifying the
chart significantly.}

\item{cor_spacing}{integer; if \code{correlations = TRUE}: width of the ring
the latent correlations between tests are drawn in; defaults to 0.4.}

\item{relative_scaling}{integer; relative size of the global chart compared to
the nested facet charts; defaults to 3.}

\item{xarrows}{data frame containing information about additional correlation
arrows between facets of different tests; see examples.}
}
\value{
List containing coordinates of chart objects.
}
\description{
Generates coordinates for a nested chart and all other charts.
}
\details{
Use this function in conjunction with \code{\link{plot_nested}} for
 nested models.

 Use \code{\link{nested_chart}} to create the chart in a single step.

 The following changes to the chart appearance are made using this function,
 not \code{\link{plot_nested}}: \itemize{ \item overall rotation \item
 rotation of nested charts \item title repositioning \item circle radius
 \item relative scaling \item size of space for correlations }

 This function also provides coordinates for facet charts and item charts (if
 \code{prepare_item_charts = TRUE}) You can use it in conjunction with
 \code{\link{plot_facets}} or \code{\link{plot_items}}.

 If you set \code{subrotate} to a single value, all nested facet charts will
 be rotated by the same amount. If you use a vector of values, the nested
 facet charts will be rotated one by one by the values from that vector.

 Increase \code{relative_scaling} to avoid circle overlap. Decrease it to
 make small chart objects more visible.

 \code{correlations} and \code{cor_spacing} add larger circles around the
 nested facet charts, but do not change these facet charts.
}
\examples{
# # creating nested charts is a two step process, using this function and
# # plot_nested:
# coord <- coord_nested(self_confidence, subradius = .6)
# sc_nested <- plot_nested(coord, filename = "sc_nested")
# sc_nested
#
# # adding xarrows
# sc_arrows <- data.frame(V1_factor = rep(NA, 3),
#                          V1_subfactor = rep(NA, 3),
#                         V2_factor = rep(NA, 3),
#                          V2_subfactor = rep(NA, 3),
#                         value = rep(NA, 3))
# sc_arrows[1, ] <- c("DSSEI", "Ab", "RSES", "Ps", ".67")
# sc_arrows[2, ] <- c("DSSEI", "Ab", "SMTQ", "Cs", ".81")
# sc_arrows[3, ] <- c("SMTQ", "Ct", "RSES", "Ns", ".76")
# coord <- coord_nested(self_confidence,
#                        subradius = .6,
#                        xarrows = sc_arrows)
# sc_nested <- plot_nested(coord, filename = "sc_nested", show_xarrows = TRUE)
# sc_nested
#
# # rotating the nested facet charts one by one
# coord <- coord_nested(self_confidence,
#                        subradius = .6,
#                        subrotate_radians = c(0, pi / 2, 0))
#  sc_nested <- plot_nested(coord, filename = "sc_nested")
#  sc_nested
}
\seealso{
\code{\link{plot_nested}} \code{\link{nested_chart}}
}
