% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_chart.R
\name{coord_nested}
\alias{coord_nested}
\title{Coord Nested}
\usage{
coord_nested(data, subradius = 0, tick = 0, rotate_radians = 0,
  rotate_degrees = 0, subrotate_radians = 0, subrotate_degrees = 0,
  rotate_construct_label_radians = 0,
  rotate_construct_label_degrees = 0, prepare_item_charts = FALSE,
  correlations = TRUE, cor_spacing = 0, relative_scaling = 0,
  xarrows = NULL)
}
\arguments{
\item{data}{list generated by \code{\link{input_excel}}, or by hand,
containing formatted data.}

\item{subradius}{integer; same unit as center distances; radius of the facet
circles; defaults to 0, in which case an appropriate value is estimated.}

\item{tick}{numeric; axis tick position; defaults to 0, in which case an
appropriate value is estimated.}

\item{rotate_radians}{integer; radian angle to rotate the chart
counter-clockwise by; use fractions of pi (e.g. pi/2 = 90 degrees).}

\item{rotate_degrees}{integer; angle in degrees to rotate the chart
counter-clockwise by.}

\item{subrotate_radians}{integer; radian angle or vector of radian angles to
rotate the nested facet charts counter-clockwise by; use fractions of pi
(e.g. pi/2 = 90 degrees).}

\item{subrotate_degrees}{integer; angle in degrees or vector of angles in
degrees to rotate the nested facet charts counter-clockwise by.}

\item{rotate_construct_label_radians}{integer; radian angle to rotate the
construct label counter-clockwise by; use fractions of pi (e.g. pi/2 = 90
degrees).}

\item{rotate_construct_label_degrees}{integer; angle in degrees to rotate the
construct label counter-clockwise by.}

\item{prepare_item_charts}{logical; if \code{TRUE}, generates the item chart
coordinates for all factors by calling \code{\link{coord_items}}.}

\item{correlations}{logical; if \code{TRUE}, generates the coordinates for the
latent correlations between tests. Sets up a ring to draw them in. If
\code{FALSE}, the ring and the correlations are omitted, simplifying the
chart significantly.}

\item{cor_spacing}{integer; if \code{correlations = TRUE}: width of the ring,
the latent correlations between tests are drawn in; defaults to 0, in which
case an appropriate value is estimated.}

\item{relative_scaling}{integer; relative size of the global chart scale
compared to the nested facet chart scales; defaults to 0, in which case an
appropriate value is estimated.}

\item{xarrows}{data frame containing information about additional correlation
arrows between facets of different tests; see examples.}
}
\value{
List containing coordinates of chart objects.
}
\description{
Generates the coordinates for a nested chart and all other charts.
}
\details{
Use \code{\link{nested_chart}} to create nested charts.
}
\seealso{
\code{\link{plot_nested}} \code{\link{nested_chart}}
}
