% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DSSEI}
\alias{DSSEI}
\title{Domain Specific Self-Esteem Inventory}
\format{An object of class \code{list} of length 3.}
\source{
Dantlgraber, M., Stieger, S., & Reips, U.-D. (in revision).
  Introducing Item Pool Visualization (IPV)
}
\usage{
DSSEI
}
\description{
A set of center distances and latent correlations for items and facets of the
DSSEI, completed by 2272 german speaking participants using the german
version of the questionnaire.
}
\keyword{datasets}
