% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_chart.R
\name{facet_chart}
\alias{facet_chart}
\title{Facet Chart}
\usage{
facet_chart(data, subradius = 0, filename = "myipv.pdf", size = 1,
  font = "sans", rotate_radians = 0, rotate_degrees = 0,
  filewidth = 10, fileheight = 10, dpi = 500, colour = "black",
  fade = 85, tick = 0, cor_labels = TRUE, rotate_title_radians = 0,
  rotate_title_degrees = 0, size_cor_labels = 1, size_test_label = 1,
  size_facet_labels = 1, width_axes = 1, width_circles = 1,
  width_tick = 1, size_tick_label = 1, size_center_dot = 1)
}
\arguments{
\item{data}{list generated by \code{\link{input_excel}}, or by hand,
containing formatted data.}

\item{subradius}{integer; same unit as center distances; radius of the
circles representing facets; use to avoid circle overlap and optimize
appearance; if 0 (default) a value is chosen automatically.}

\item{filename}{character; name of the file to save. Supported formats are:
"pdf" (best quality with lowest file size), "png", "jpeg". Use "none" to
suppress file output; defaults to "myipv.pdf".}

\item{size}{integer; smartly scaled size of chart objects.}

\item{font}{character; text font; defaults to 'sans' (use extrafonts to
install additional fonts).}

\item{rotate_radians}{integer; radian angle to rotate the chart
counter-clockwise by; use fractions of pi (e.g. pi/2 = 90 degrees).}

\item{rotate_degrees}{integer; angle in degrees to rotate the chart
counter-clockwise by.}

\item{filewidth}{integer; width of the graphic in inches; defaults to 10.}

\item{fileheight}{integer; height of the graphic in inches; defaults to 10.}

\item{dpi}{integer; resolution in dots per inch; defaults to 500.}

\item{colour}{character; name of the accent colour.}

\item{fade}{integer; brightness of the gray tones between 0 (black) and 100
(white) in steps of 1; defaults to 85.}

\item{tick}{numeric; axis tick position; if 0 (default) a value is chosen
automatically.}

\item{cor_labels}{logical; if \code{TRUE}, draws latent correlations between
facets as text.}

\item{rotate_title_radians}{integer; radian angle to rotate the global label
counter-clockwise by; use fractions of pi (e.g. pi/2 = 90 degrees).}

\item{rotate_title_degrees}{integer; angle in degrees to rotate the global
label counter-clockwise by.}

\item{size_cor_labels}{integer; font size of the latent correlations between
facets (relative to default).}

\item{size_test_label}{integer; font size of the test label (relative to
default).}

\item{size_facet_labels}{integer; font size of the facet labels (relative to
default).}

\item{width_axes}{integer; width of the radial axis lines indicating center
distances (relative to default).}

\item{width_circles}{integer; width of the circle outlines (relative to
default).}

\item{width_tick}{integer; width of the dotted axis tick line (relative to
default).}

\item{size_tick_label}{integer; font size of the axis tick label (relative to
default).}

\item{size_center_dot}{integer; size of the center dot marking the origin
(relative to default).}
}
\value{
Object of the class "ggplot" and optionally the same object saved as
  a file.
}
\description{
Creates a facet chart, showing the facets of a test.
}
\examples{
# as simple as that
facet_chart(SMTQ, subradius = .6)

}
\seealso{
\code{\link{item_chart}} \code{\link{nested_chart}}
}
