% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv.R
\name{Fusion}
\alias{Fusion}
\title{Fusion}
\usage{
Fusion(X, Y, varX, varY, canon, transX = NULL, transY = NULL)
}
\arguments{
\item{X}{First dataframe to merge}

\item{Y}{Second dataframe to merge}

\item{varX}{Variable acting as key in X}

\item{varY}{Variable acting as key in Y}

\item{canon}{Correspondence between varX, varY and canonical value of the key}

\item{transX}{Transformation to be applied to varX (if any)}

\item{transY}{Transformation to be applied to varY (if any)}
}
\value{
Merged data frame (or lines in X, y or both that cannot be matched)
}
\description{
Merge two dataframes using  \code{varX} and \code{varY} as keys. Since they may not be exactly alike,
\code{canon} contains a dataframe which enables translation of either key
to a canonical form after (optionally) transliterating to a common character encoding.
}
\examples{
X     <- data.frame(a=c("Bilbao","San Sebastian","Vitoria","Terue"),
                    b=c("A","B","B","D"), d=c(TRUE,TRUE,FALSE,TRUE))
Y     <- data.frame(e=c("Bilbao","Donostia","Vitoria/Gazteiz","Sorias"),
                    g=c("F","G","H","J"), h=c(TRUE,FALSE,TRUE,TRUE))
canon <- data.frame(X=c("Bilbao","San Sebastian","Vitoria"),
                    Y=c("Bilbao","Donostia","Vitoria/Gazteiz"),
                    Canon=c("Bilbao","Donostia","Vitoria"))
XY <- Fusion(X,Y,varX="a",varY="e",canon=canon)

canon <- data.frame(X=c("Bilbao","San Sebastian","Vitoria","Teruel"),
                    Y=c("Bilbao","Donostia","Vitoria/Gazteiz", "Teruel"),
                    Canon=c("Bilbao","Donostia","Vitoria","Teruel"))
XY <- Fusion(X,Y,varX="a",varY="e",canon=canon)

canon <- data.frame(X=c("Bilbao","San Sebastian","Vitoria","Teruel","Soria"),
                    Y=c("Bilbao","Donostia","Vitoria/Gazteiz", "Teruel","Soria"),
                    Canon=c("Bilbao","Donostia","Vitoria","Teruel","Soria"))
XY <- Fusion(X,Y,varX="a",varY="e",canon=canon)

}
