% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv.R
\name{IndZonas}
\alias{IndZonas}
\title{IndZonas}
\usage{
IndZonas(datos, zonas, frm = NULL, base = "2008-01-01", plot = FALSE)
}
\arguments{
\item{datos}{Dataframe of observations}

\item{zonas}{Name of the column of \code{datos} defining the areeas for which we want indices.}

\item{frm}{Formula specifying the base model, in the notation expected by  \code{gam.} The formula is common to all indices computed byu the function.}

\item{base}{Base date (index=100). It is a day in format yyyy-mm-dd.}

\item{plot}{\code{FALSE} (default) if no graphs are desired, \code{TRUE} otherwise.}
}
\value{
A list with the computed indices, in the format in which they are returned by \code{ConsInd}.
}
\description{
Estimation of global indices by areas; merely a wrapper around  \code{ConsInd}.
}
\examples{

}
