% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_excel.R
\name{input_excel}
\alias{input_excel}
\title{Input Excel}
\usage{
input_excel(global = NULL, factors)
}
\arguments{
\item{global}{character; name of the excel file containing data on the overall general factor and the factors; see examples.}

\item{factors}{character; name(s) of the excel file(s) containing data on the factor(s) and their/its facets; see examples.}
}
\value{
list containing formatted data including center distances for use with \code{\link{model_items}}, \code{\link{model_facets}}, and \code{\link{model_nested}}.
}
\description{
This function reads excel files containing factor loadings and latent correlations for IPV plots.
}
\details{
Note that the excel files need a very specific structure, see examples.
The \code{global} argument defaults to NULL.
This allows to read data on a single factor and its facets to the \code{factors} argument.
The result is a more simple model that can be displayed by an item plot and a facet plot.
}
