% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_excel_factor.R
\name{input_excel_factor}
\alias{input_excel_factor}
\title{Input Excel Factor}
\usage{
input_excel_factor(file)
}
\arguments{
\item{file}{character; filename of the excel file}
}
\value{
list containing formatted data including center distances for \code{\link{model_items}}, \code{\link{model_facets}}
}
\description{
Reads factor loadings and latent correlations from an excel file.
}
\details{
Helper function of \code{\link{input_excel}}.
}
\seealso{
\code{\link{input_excel}}
}
