% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPV.R
\docType{package}
\name{IPV}
\alias{IPV}
\alias{IPV-package}
\title{IPV: A package to create Item Pool Visualizations}
\description{
The IPV package provides three categories of important functions:
input, model, and plot.
}
\section{Input Functions}{


The input functions format your raw data, so the model functions can use them.
\code{\link{input_excel}}
}

\section{Model Functions}{


The model functions calculate specific coordinates the plot functions draw from.
\code{\link{model_items}} \code{\link{model_facets}} \code{\link{model_nested}}
}

\section{Plot Functions}{


The plot functions create and customize the actual plots.
\code{\link{plot_items}} \code{\link{plot_facets}} \code{\link{plot_nested}}
}

