% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPV.R
\docType{package}
\name{IPV}
\alias{IPV}
\alias{IPV-package}
\title{IPV: A package to create Item Pool Visualizations}
\description{
The IPV package provides two categories of important functions: input, and
chart.
}
\section{Input Functions}{


  The input functions help preprocessing your data, so the other functions
  can use them. \code{\link{input_manual_simple}}
  \code{\link{input_manual_nested}} \code{\link{input_manual_process}}
  \code{\link{input_excel}}
}

\section{Chart Functions}{


  Chart functions create a ggplot2 object (the chart) and optionally save
  this object to a file. There are three types of charts.
  \code{\link{item_chart}} \code{\link{facet_chart}}
  \code{\link{nested_chart}}
}

\section{Workflow}{


  Prepare your data using input functions or by hand. Choose a chart type and
  use the proper chart function. Then \enumerate{\item choose parameter
  values of the chart function \item check chart appearance \item repeat
  until satisfied}
}

