% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_chart.R
\name{item_chart}
\alias{item_chart}
\title{Item Chart}
\usage{
item_chart(data, rotate_radians = 0, rotate_degrees = 0, size = 1,
  file = "pdf", filename = NULL, filewidth = 10, fileheight = 10,
  colour = "black", colour2 = "black", fade = 90, font = "sans",
  tick_label = TRUE, size_tick_label = 1, size_title = 1,
  size_facet_labels = 1, width_axes = 1, size_arrow_heads = 1,
  width_items = 1, width_grid = 1, size_center_dot = 1)
}
\arguments{
\item{data}{list generated by \code{\link{input_excel}}, or by hand,
containing formatted data.}

\item{rotate_radians}{integer; radian angle to rotate the chart
counter-clockwise by; use fractions of pi (e.g. pi/2 = 90 degrees).}

\item{rotate_degrees}{integer; angle in degrees to rotate the chart
counter-clockwise by.}

\item{size}{integer; smartly scaled size of chart objects.}

\item{file}{character; file format to save. Supported formats are: "pdf"
(best quality with lowest file size),"png","jpeg". Suppress file output
using any other value; defaults to "pdf".}

\item{filename}{character string; name of the file the chart is written to;
filename extension added from file parameter like this: [filename].[file].}

\item{filewidth}{integer; width of the .pdf document; defaults to 10; dpi is
3000.}

\item{fileheight}{integer; height of the .pdf document; defaults to 10, dpi
is 3000.}

\item{colour}{character; name of the first accent colour.}

\item{colour2}{character; name of the second accent colour.}

\item{fade}{integer; brightness of minor grid lines, gray tones between 0
(black) and 100 (white) in steps of 1; defaults to 90.}

\item{font}{character; text font; defaults to 'sans' (use extrafonts to
install additional fonts).}

\item{tick_label}{logical; if \code{TRUE}, draws a text label for the axis
tick}

\item{size_tick_label}{integer; font size of the axis tick label (relative to
default).}

\item{size_title}{integer; font size of the test label (relative to default).}

\item{size_facet_labels}{integer; font size of the axis labels (relative to
default).}

\item{width_axes}{integer; width of the radial axis lines indicating center
distances (relative to default).}

\item{size_arrow_heads}{integer; length of the arrow heads at the end of the
axes (relative to default).}

\item{width_items}{integer; width of the item bars (relative to default).
Currently not supported!}

\item{width_grid}{integer; width of the dotted grid lines (relative to
default).}

\item{size_center_dot}{integer; size of the center dot marking the origin
(relative to default).}
}
\value{
Object of the class "ggplot" and optionally the same object saved as
  a file.
}
\description{
Creates an item chart, showing the items of a test arranged by facets.
}
\details{
This function creates the chart in one step, but does not return the
  coordinates from \code{\link{coord_items}}. Alternatively use
  \code{\link{coord_items}} and \code{\link{plot_items}}. This allows you to
  make changes to or reuse the coordinates.
}
\examples{
# as simple as that
item_chart(SMTQ)

}
\seealso{
\code{\link{coord_items}} \code{\link{plot_items}}
}
