% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_facets.R
\name{model_facets}
\alias{model_facets}
\title{Model Facets}
\usage{
model_facets(data, subradius, rotate = 0)
}
\arguments{
\item{data}{list generated by \code{\link{input_excel}} containing formatted data.}

\item{subradius}{integer; radius of the circles representing facets (same unit as center distances).}

\item{rotate}{integer; radian angle to rotate the plot counter-clockwise by (between 0 and 2*pi).}
}
\value{
A list of dataframes containing the coordinates of the plot objects.
}
\description{
This function generates the coordinates for a facet plot.
}
\details{
Use this function to be able to create a plot with \code{\link{plot_facets}}
using data you read from an excel file with \code{\link{input_excel}}.
You can also create \code{data} yourself, but \code{data} needs the same structure as if
generated by \code{\link{input_excel}}.
If \code{\link{input_excel}} generated lists for multiple factors, only use one.
}
