% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_items.R
\name{model_items}
\alias{model_items}
\title{Model Items}
\usage{
model_items(data, rotate = 0)
}
\arguments{
\item{data}{list generated by list generated by \code{\link{input_excel}} containing formatted data.}

\item{rotate}{integer; radian angle to rotate the plot counter-clockwise by (between 0 and 2*pi).}
}
\value{
A list of dataframes containing the coordinates of the plot objects.
}
\description{
This function generates the coordinates for an item plot.
}
\details{
Use this function to be able to create a plot with \code{\link{plot_items}}
using data you read from an excel file with \code{\link{input_excel}}.
You can also create \code{data} yourself, but \code{data} needs the same structure as if
generated by \code{\link{input_excel}}.
If \code{\link{input_excel}} generated lists for multiple factors, only use one.
}
\examples{
# creating plots is a two step process, using model_ and plot_ functions:
coord <- model_items(DSSEI)
DSSEI_item_plot <- plot_items(coord,filename = "DSSEI_items")

}
\seealso{
\code{\link{plot_items}}
}
