% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_nested.R
\name{model_nested}
\alias{model_nested}
\title{Model Nested}
\usage{
model_nested(data, subradius, rotate = 0, subrotate = 0, rotate_title = 0,
  items = FALSE, cors = TRUE, cor_spacing = 0.3, relative_scaling = 1.5,
  extra_arrows = NULL)
}
\arguments{
\item{data}{generated by \code{\link{input_excel}} containing formatted data.}

\item{subradius}{integer; radius of the circles representing facets (same unit as center distances).}

\item{rotate}{integer; radian angle to rotate the plot counter-clockwise by (between 0 and 2*pi).}

\item{subrotate}{integer; radian angle or vector of radian angles to rotate the nested plots counter-clockwise by (between 0 and 2*pi).}

\item{rotate_title}{integer; radian angle to rotate the overall general factor label counter-clockwise by (between 0 and 2*pi).}

\item{items}{logical; if TRUE, generates the coordinates for item plots for all factors using \code{\link{model_items}}.}

\item{cors}{logical; if TRUE, generates coordinates for latent correlations between factors and sets up a ring to plot them in.}

\item{cor_spacing}{integer; width of the ring the latent correlations between factors are plotted in; defaults to 0.3.}

\item{relative_scaling}{integer; relative size of the global to the nested plots; defaults to 1.5.}

\item{extra_arrows}{a data frame containing information about additional correlation arrows between facets of different factors; see examples.}
}
\value{
A list of lists of dataframes containing the coordinates of the plot objects.
}
\description{
This function generates the coordinates for a nested plot and all other plots.
}
\details{
Use this function to be able to create a plot with \code{\link{plot_nested}}
using data you read from an excel file with \code{\link{input_excel}}.
You can also create \code{data} yourself, but \code{data} needs the same structure as if
generated by \code{\link{input_excel}}.
This function can also provide data for facet plots and item plots
you can create using \code{\link{plot_facets}} and \code{\link{plot_items}}.
The \code{subrotate} argument can be a single value. In this case, all nested plots will be rotated in the same way.
It can also be a vector of values of the same length as the number of nested plots within the global plot. In this
case, every nested plot will be rotated by its specific value given in the vector.
Increase the value of the \code{relative_scaling} argument to avoid circle overlap. Decrease it to make small circles
more visible.
The \code{cor_spacing} argument increases the size of the nested plot's circles, keeping the scale of its inner parts as is.
}
