% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_facets.R
\name{plot_facets}
\alias{plot_facets}
\title{Plot Facets}
\usage{
plot_facets(coor, size = 1, filename, width = 10, height = 10,
  colour = "black", cor_labels = TRUE, tick = 0.1, font = "Helvetica",
  size_title = 1, size_subfactor_labels = 1, width_axes = 1,
  width_circles = 1, size_tick = 1, size_tick_label = 1,
  size_cor_labels = 1, size_center_dot = 1)
}
\arguments{
\item{coor}{list generated by \code{\link{model_facets}} or \code{\link{model_nested}} containing the coordinates of the parts of the plot.}

\item{size}{integer; smartly scaled size of all parts of the plot.}

\item{filename}{character string; name of the .pdf file the plot is written to; filename extension adds automatically.}

\item{width}{integer; width of the .pdf file; defaults to 10.}

\item{height}{integer; height of the .pdf file; defaults to 10.}

\item{colour}{character; name of the accent colour.}

\item{cor_labels}{logical; if TRUE, plots latent correlations between facets as text.}

\item{tick}{numeric; axis tick position; defaults to .1.}

\item{font}{character; text font; defaults to 'Helvetica' (use extrafonts to install fonts you do not have).}

\item{size_title}{integer; font size of the factor label (relative to default).}

\item{size_subfactor_labels}{integer; font size of the subfactor label (relative to default).}

\item{width_axes}{integer; width of the radial axis lines (relative to default).}

\item{width_circles}{integer; width of the circle outlines (relative to default).}

\item{size_tick}{integer; width of the dotted axis tick line (relative to default).}

\item{size_tick_label}{integer; font size of the axis tick label (relative to default).}

\item{size_cor_labels}{integer; font size of the latent correlations between facets (relative to default).}

\item{size_center_dot}{integer; size of the center dot marking the origin (relative to default).}
}
\value{
an object of the class "ggplot" and the same object as a .pdf file.
}
\description{
This function plots a factor and its facets as circles.
}
\details{
Facets are displayed using the average center distance of their items
as distance from the origin to the circle's edge.
}
\examples{
# creating plots is a two step process, using model_ and plot_ functions:
coord <- model_facets(DSSEI,subradius = .5)
DSSEI_item_plot <- plot_facets(coord,filename = "DSSEI_facets")

}
\seealso{
\code{\link{model_facets}}
}
