% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_facets.R
\name{plot_facets}
\alias{plot_facets}
\title{Plot Facets}
\usage{
plot_facets(coor, size = 1, filesave = TRUE, filename = NULL,
  filewidth = 10, fileheight = 10, colour = "black", cor_labels = TRUE,
  tick = 0.1, font = "sans", size_test_label = 1, size_facet_labels = 1,
  width_axes = 1, width_circles = 1, size_tick = 1, size_tick_label = 1,
  size_cor_labels = 1, size_center_dot = 1)
}
\arguments{
\item{coor}{list generated by \code{\link{coord_facets}} or
\code{\link{coord_nested}}.}

\item{size}{integer; smartly scaled size of chart objects.}

\item{filesave}{logical, if TRUE, saves the chart to the working directory as
[filename].pdf}

\item{filename}{character string; name of the .pdf file the chart is written
to; filename extension added automatically: [filename].pdf.}

\item{filewidth}{integer; width of the .pdf file; defaults to 10, dpi is 3000.}

\item{fileheight}{integer; height of the .pdf file; defaults to 10, dpi is 3000.}

\item{colour}{character; name of the accent colour.}

\item{cor_labels}{logical; if \code{TRUE}, draws latent correlations between
facets as text.}

\item{tick}{numeric; axis tick position; defaults to .1.}

\item{font}{character; text font; defaults to 'sans' (use extrafonts to
install additional fonts).}

\item{size_test_label}{integer; font size of the test label (relative to
default).}

\item{size_facet_labels}{integer; font size of the facet labels
(relative to default).}

\item{width_axes}{integer; width of the radial axis lines (relative to
default).}

\item{width_circles}{integer; width of the circle outlines (relative to
default).}

\item{size_tick}{integer; width of the dotted axis tick line (relative to
default).}

\item{size_tick_label}{integer; font size of the axis tick label (relative to
default).}

\item{size_cor_labels}{integer; font size of the latent correlations between
facets (relative to default).}

\item{size_center_dot}{integer; size of the center dot marking the origin
(relative to default).}
}
\value{
Object of the class "ggplot" and the same object as a .pdf file.
}
\description{
Plots a facet chart, showing the facets of a test.
}
\details{
Use this function in conjunction with \code{\link{coord_facets}} or
  \code{\link{coord_nested}} for simple models.

  Facets are displayed using the average center distance of their items as
  distance from the origin to the circle's edge.
}
\examples{
# creating facet charts is a two step process, using coord_facets and this function:
coord <- coord_facets(DSSEI,subradius = .5)
DSSEI_facets <- plot_facets(coord,filename = "DSSEI_facets")
DSSEI_facets

}
\seealso{
\code{\link{coord_facets}}
}
