% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_facets.R
\name{plot_facets}
\alias{plot_facets}
\title{Plot Facets}
\usage{
plot_facets(coord, size = 1, filename = "myipv.pdf", filewidth = 10,
  fileheight = 10, colour = "black", fade = 90, font = "sans",
  tick = 0.1, cor_labels = TRUE, size_cor_labels = 1,
  size_test_label = 1, size_facet_labels = 1, width_axes = 1,
  width_circles = 1, width_tick = 1, size_tick_label = 1,
  size_center_dot = 1)
}
\arguments{
\item{coord}{list generated by \code{\link{coord_facets}} or
\code{\link{coord_nested}}.}

\item{size}{integer; smartly scaled size of chart objects.}

\item{filename}{character; name of the file to save. Supported formats are:
"pdf" (best quality with lowest file size), "png", "jpeg". Use "none" to
suppress file output; defaults to "myipv.pdf".}

\item{filewidth}{integer; width of the .pdf file; defaults to 10, dpi is
3000.}

\item{fileheight}{integer; height of the .pdf file; defaults to 10, dpi is
3000.}

\item{colour}{character; name of the accent colour.}

\item{fade}{integer; brightness of the gray tones between 0 (black) and 100
(white) in steps of 1; defaults to 90.}

\item{font}{character; text font; defaults to 'sans' (use extrafonts to
install additional fonts).}

\item{tick}{numeric; axis tick position; defaults to .1.}

\item{cor_labels}{logical; if \code{TRUE}, draws latent correlations between
facets as text.}

\item{size_cor_labels}{integer; font size of the latent correlations between
facets (relative to default).}

\item{size_test_label}{integer; font size of the test label (relative to
default).}

\item{size_facet_labels}{integer; font size of the facet labels (relative to
default).}

\item{width_axes}{integer; width of the radial axis lines indicating center
distances (relative to default).}

\item{width_circles}{integer; width of the circle outlines (relative to
default).}

\item{width_tick}{integer; width of the dotted axis tick line (relative to
default).}

\item{size_tick_label}{integer; font size of the axis tick label (relative to
default).}

\item{size_center_dot}{integer; size of the center dot marking the origin
(relative to default).}
}
\value{
Object of the class "ggplot" and optionally the same object saved as
  a file.
}
\description{
Plots a facet chart, showing the facets of a test.
}
\details{
Use this function in conjunction with \code{\link{coord_facets}} or
  \code{\link{coord_nested}} for simple models.

  Use \code{\link{facet_chart}} to create the chart in a single step.

  Rotate  or change the size of the circles using \code{\link{coord_facets}},
  not this function.

  When changing the size of objects, consider the \code{size} parameter first
  and make specific adjustments with the other \code{size_} and \code{width_}
  parameters after.

  Facets are displayed using the average center distance of their items as
  distance from the origin to the circle's edge.
}
\examples{
# # creating facet charts is a two step process, using coord_facets and this
# # function:
# coord <- coord_facets(DSSEI, subradius = .5)
# DSSEI_facets <- plot_facets(coord, filename = "DSSEI_facets.pdf")
# DSSEI_facets
}
\seealso{
\code{\link{coord_facets}} \code{\link{facet_chart}}
}
