% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_items.R
\name{plot_items}
\alias{plot_items}
\title{Plot Items}
\usage{
plot_items(coord, size = 1, filename = "myipv.pdf", filewidth = 12,
  fileheight = 10, dpi = 500, color = "black", color2 = "black",
  fade = 0, font = "sans", size_tick_label = 1, size_title = 1,
  size_facet_labels = 1, width_axes = 1, size_arrow_heads = 1,
  width_items = 1, width_grid = 1)
}
\arguments{
\item{coord}{list generated by \code{\link{coord_items}} or
\code{\link{coord_nested}}.}

\item{size}{integer; size of all chart objects except item bars.}

\item{filename}{character; name of the file to save. Supported formats are:
"pdf" (best quality with lowest file size), "png", "jpeg". Use "none" to
suppress file output; defaults to "myipv.pdf".}

\item{filewidth}{integer; width of the .pdf document; defaults to 12.}

\item{fileheight}{integer; height of the .pdf document; defaults to 10.}

\item{dpi}{integer; resolution in dots per inch; defaults to 500.}

\item{color}{character; name of the first accent color; defaults to black.}

\item{color2}{character; name of the second accent color; defaults to
black.}

\item{fade}{integer; brightness of grid lines, axes, and center dot; values
can range from -4 (almost white) to 2.6 (almost black) in steps of .2;
defaults to 0.}

\item{font}{character; text font; defaults to 'sans' (use extrafonts to
install additional fonts).}

\item{size_tick_label}{integer; font size of the axis tick label (relative to
default).}

\item{size_title}{integer; font size of the test label (relative to default).}

\item{size_facet_labels}{integer; font size of the axis labels (relative to
default).}

\item{width_axes}{integer; width of the radial axis lines indicating center
distances (relative to default).}

\item{size_arrow_heads}{integer; length of the arrow heads at the end of the
axes (relative to default).}

\item{width_items}{integer; width of the item bars (relative to default).}

\item{width_grid}{integer; width of the dotted grid lines (relative to
default).}
}
\value{
Object of the class "ggplot" and optionally the same object saved as
  a file.
}
\description{
Plots an item chart, showing the items of a test arranged by facets.
}
\details{
Use this function in conjunction with \code{\link{coord_items}} or
  \code{\link{coord_nested}} for simple models.

  Use \code{\link{item_chart}} to create the chart in a single step.

  Rotate using \code{\link{coord_items}}, not this function.

  When changing the size of objects, consider the \code{size} parameter first
  and make specific adjustments with the other \code{size_} and \code{width_}
  parameters after.

  To better display overlapping item values, change the width of the item
  bars, or set the accent colors to different values, or change the ratio of
  item lengths.
}
\examples{
# # creating item charts is a two step process, using coord_items and this
# # function:
# coord <- coord_items(DSSEI)
# DSSEI_items <- plot_items(coord, filename = "DSSEI_items.pdf")
# DSSEI_items
}
\seealso{
\code{\link{coord_items}} \code{\link{item_chart}}
}
