% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_items.R
\name{plot_items}
\alias{plot_items}
\title{Plot Items}
\usage{
plot_items(coor, size = 1, filename, width = 10, height = 10,
  font = "Helvetica", colour = "black", colour2 = "black",
  tick_label = TRUE, size_title = 1, size_axis_labels = 1,
  width_axes = 1, size_arrow_heads = 1, width_items = 1, size_grid = 1,
  size_tick_label = 1, size_center_dot = 1)
}
\arguments{
\item{coor}{list generated by \code{\link{model_items}} or \code{\link{model_nested}} containing the coordinates of the parts of the plot.}

\item{size}{integer; smartly scaled size of all parts of the plot.}

\item{filename}{character string; name of the .pdf file the plot is written to; filename extension adds automatically.}

\item{width}{integer; width of the .pdf file; defaults to 10.}

\item{height}{integer; height of the .pdf file; defaults to 10.}

\item{font}{character; text font; defaults to 'Helvetica'.}

\item{colour}{character; name of the accent colour.}

\item{colour2}{character; name of the second accent colour used to better distinguish between similar items}

\item{tick_label}{logical; if TRUE, plots a text label for the axis tick}

\item{size_title}{integer; font size of the factor label (relative to default).}

\item{size_axis_labels}{integer; font size of the axis labels (relative to default).}

\item{width_axes}{integer; width of the radial axis lines (relative to default).}

\item{size_arrow_heads}{integer; length of the arrow heads (relative to default).}

\item{width_items}{integer; width of the item bars (relative to default). Currently not supported!}

\item{size_grid}{integer; width of the dotted grid lines (relative to default).}

\item{size_tick_label}{integer; font size of the axis tick label (relative to default).}

\item{size_center_dot}{integer; size of the center dot marking the origin (relative to default).}
}
\value{
an object of the class "ggplot" and the same object as a .pdf file.
}
\description{
This function plots a factor and its items as bars on radial axes representing its facets.
}
