% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nested.R
\name{plot_nested}
\alias{plot_nested}
\title{Plot Nested}
\usage{
plot_nested(coord, size = 1, file = "pdf", filename = NULL,
  filewidth = 10, fileheight = 10, cor_labels_tests = TRUE,
  cor_labels_facets = TRUE, colour_tests = "black",
  colour_facets = "black", fade = 90, font = "sans",
  extra_arrows = FALSE, tick = 0.1, size_title = 1,
  size_test_labels = 1, size_facet_labels = 1, width_axes = 1,
  width_axes_inner = 1, width_circles = 1, width_circles_inner = 1,
  size_tick = 1, size_tick_inner = 1, size_tick_label = 1,
  size_cor_labels = 1, size_cor_labels_inner = 1, size_center_dot = 1,
  size_center_dot_inner = 1, size_extra_arrows = 1,
  size_extra_arrow_heads = 1, size_extra_labels = 1)
}
\arguments{
\item{coord}{list generated by \code{\link{coord_nested}}.}

\item{size}{integer; smartly scaled size of chart objects.}

\item{file}{character; file format to save. Supported formats are: "pdf"
(best quality with lowest file size),"png","jpeg". Suppress file output using any
other value; defaults to "pdf".}

\item{filename}{character string; name of the file the chart is written to;
filename extension added from file parameter like this: [filename].[file].}

\item{filewidth}{integer; width of the .pdf file; defaults to 10, dpi is 3000.}

\item{fileheight}{integer; height of the .pdf file; defaults to 10, dpi is 3000.}

\item{cor_labels_tests}{logical; if \code{TRUE}, draws latent correlations between
tests as text.}

\item{cor_labels_facets}{logical; if \code{TRUE}, draws latent
correlations between facets as text.}

\item{colour_tests}{character; name of the global accent colour.}

\item{colour_facets}{character; name of the nested accent colour.}

\item{fade}{integer; brightness of the gray tones between 0 (black) and 100
(white) in steps of 1; defaults to 90.}

\item{font}{character; text font; defaults to 'sans' (use extrafonts to
install additional fonts).}

\item{extra_arrows}{logical; if \code{TRUE}, draws correlation arrows between
facets of different tests.}

\item{tick}{numeric; axis tick position; defaults to .1.}

\item{size_title}{integer; font size of the global label (relative to
default).}

\item{size_test_labels}{integer; font size of the test labels (relative
to default).}

\item{size_facet_labels}{integer; font size of the facet labels
(relative to default).}

\item{width_axes}{integer; width of the global radial axis lines (relative to
default).}

\item{width_axes_inner}{integer; width of the nested radial axis lines
(relative to default).}

\item{width_circles}{integer; width of the global circle outlines (relative
to default).}

\item{width_circles_inner}{integer; width of the nested circle outlines
(relative to default).}

\item{size_tick}{integer; width of the global dotted axis tick line (relative
to default).}

\item{size_tick_inner}{integer; width of the nested dotted axis tick line
(relative to default).}

\item{size_tick_label}{integer; font size of the axis tick label (relative to
default).}

\item{size_cor_labels}{integer; font size of the latent correlations between
tests (relative to default).}

\item{size_cor_labels_inner}{integer; font size of the latent correlations
between facets (relative to default).}

\item{size_center_dot}{integer; size of the center dot marking the global
origin (relative to default).}

\item{size_center_dot_inner}{integer; size of the center dots marking the
nested origins (relative to default).}

\item{size_extra_arrows}{integer; width of the extra arrow lines (relative to
default).}

\item{size_extra_arrow_heads}{integer; length of the extra arrow heads
(relative to default).}

\item{size_extra_labels}{integer; font size of the latent correlations
indicated by extra arrows (relative to default).}
}
\value{
Object of the class "ggplot" and the same object as a .pdf file.
}
\description{
Plots a nested chart, showing several tests and their facets in a global
chart.
}
\details{
Use this function in conjunction with \code{\link{coord_nested}} for
  nested models.

  The nested chart treats tests like facets of a global factor. It uses facet
  charts for tests similar to how facet charts use circles for facets.

  Rotate by changing your \code{\link{coord_nested}} function call.

  To add extra_arrows, you need to provide additional information to
  \code{\link{coord_nested}}.
}
\examples{
# creating nested charts is a two step process, using coord_nested and this function:
coord <- coord_nested(self_confidence, subradius = .6)
sc_nested <- plot_nested(coord, filename = "sc_nested")
sc_nested

# adding extra arrows
sc_arrows <- data.frame(V1_factor=rep(NA,3),
                        V1_subfactor=rep(NA,3),
                        V2_factor=rep(NA,3),
                        V2_subfactor=rep(NA,3),
                        value=rep(NA,3))
sc_arrows[1,] <- c("DSSEI", "Ab", "RSES", "Ps", ".67")
sc_arrows[2,] <- c("DSSEI", "Ab", "SMTQ", "Cs", ".81")
sc_arrows[3,] <- c("SMTQ", "Ct", "RSES", "Ns", ".76")
coord <- coord_nested(self_confidence, subradius = .6, extra_arrows = sc_arrows)
sc_nested <- plot_nested(coord, filename = "sc_nested", extra_arrows = TRUE)
sc_nested

}
\seealso{
\code{\link{coord_nested}}
}
