% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_factor.R
\name{shift_factor}
\alias{shift_factor}
\title{Shift Factor}
\usage{
shift_factor(data, x, y)
}
\arguments{
\item{data}{list generated by \code{\link{coord_facets}}.}

\item{x}{integer; horizontal shift.}

\item{y}{integer; vertical shift.}
}
\value{
List containing a limited selection of coordinates.
}
\description{
Moves coordinates given by \code{\link{coord_facets}} to position nested
facet graphics in \code{\link{coord_nested}}.
}
\details{
Helper function of \code{\link{coord_nested}}.
}
