% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SMTQ}
\alias{SMTQ}
\title{Sports Mental Toughness Questionnaire}
\format{An object of class \code{list} of length 2.}
\source{
Dantlgraber, M., Stieger, S., & Reips, U.-D. (in revision).
  Introducing Item Pool Visualization (IPV)
}
\usage{
SMTQ
}
\description{
A set of center distances and latent correlations for items and facets of the
SMTQ, completed by 2272 german speaking participants using the german version
of the questionnaire.
}
\section{Items}{


  __ Confidence (Cf) __

  13. I interpret potential threats as positive opportunities

  5.  I have an unshakeable confidence in my ability

  11. I have qualities that set me apart from other competitors

  6.  I have what it takes to perform well while under pressure

  14. Under pressure, I am able to make decisions with confidence and
  commitment

  1.  I can regain my composure if I have momentarily lost it

  __ Constancy (Cs) __

  3.  I am committed to completing the tasks I have to do

  12. I take responsibility for setting myself challenging targets

  8.  I give up in difficult situations

  10. I get distracted easily and lose my concentration

  __ Control (Ct) __

  2.  I worry about performing poorly

  4.  I am overcome by self-doubt

  9.  I get anxious by events I did not expect or cannot control

  7.  I get angry and frustrated when things do not go my way

  (Sheard, M., Golby, J., & Van Wersch, A. (2009). Progress toward construct
  validation of the Sports Mental Toughness Questionnaire (SMTQ). European
  Journal of Psychological Assessment, 25(3), 186-193.)
}

\keyword{datasets}
